+++ udata_hydra/routes/checks.py
@@ -64,10 +64,8 @@ async def create_check(request: web.Request) -> web.Response:
     try:
         payload: dict = await request.json()
         resource_id: str = payload["resource_id"]
-    except ValidationError as err:
-        raise web.HTTPBadRequest(text=json.dumps(err.messages))
-    except KeyError as e:
-        raise web.HTTPBadRequest(text=f"Missing key: {str(e)}")
+    except Exception as err:
+        raise web.HTTPBadRequest(text=json.dumps({"error": str(err)}))
 
     # Get URL from resource_id
     try: