+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Improve timing of checks depending on resources last modification date [#163](https://github.com/datagouv/hydra/pull/163)
 
 ## 2.0.3 (2024-10-22)
 
@@ -27,7 +27,6 @@
 - Add a `quiet` argument for `purge_check` and `purge_csv_table` CLIs [#184](https://github.com/datagouv/hydra/pull/184)
 - Fix wrong resource status [#187](https://github.com/datagouv/hydra/pull/187)
 - More informative error relative to check resource CLI [#188](https://github.com/datagouv/hydra/pull/188)
-- Improve timing of checks depending on resources last modification date [#163](https://github.com/datagouv/hydra/pull/163)
 
 ## 2.0.0 (2024-09-24)