+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Current (in progress)
 
 - Fix wrong resource status [#196](https://github.com/datagouv/hydra/pull/196)
+- Fix issue related to empty `table_indexes` column instead of default `{}` [#197](https://github.com/datagouv/hydra/pull/197)
+
 
 ## 2.0.3 (2024-10-22)